# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Jobify",
        "description": "Thông tin dự án Jobify",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8040/api",
            "description": "Local server"
        }
    ],
    "paths": {
...
        "/profile/current-user": {
            "get": {
                "tags": [
                    "PersonalInfo"
                ],
                "summary": "Get Profile Current User",
                "operationId": "f857d95767a80bbc1860351698e7fdc1",
                "responses": {
                    "200": {
                        "description": "Get user info successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Get user info successfully"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "The user is not logged in",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "The user is not logged in"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Get user info failed!",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Get user info failed!"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/update-personal-info": {
            "post": {
                "tags": [
                    "PersonalInfo"
                ],
                "summary": "Update Personal Profile",
                "description": "Update Personal Profile",
                "operationId": "c735594815021a96f3ed5614eac0310d",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "full_name",
                                    "phone_number",
                                    "position",
                                    "gender_id",
                                    "birth_date"
                                ],
                                "properties": {
                                    "full_name": {
                                        "description": "Nhập họ và tên",
                                        "type": "string",
                                        "example": "Nguyễn Văn A"
                                    },
                                    "phone_number": {
                                        "description": "Nhập số điện thoại vào nhé",
                                        "type": "string",
                                        "example": "0912345678"
                                    },
                                    "position": {
                                        "description": "Chức vụ",
                                        "type": "string",
                                        "example": "Backend Developer"
                                    },
                                    "gender_id": {
                                        "description": "Mã giới tính: 1:male, 2:female, 3:other",
                                        "type": "integer",
                                        "example": "1"
                                    },
                                    "birth_date": {
                                        "description": "\n     *                          Ngày sinh có định dạng YYYY-MM-DD, có thể không cần gửi 0 trước ngày vàng tháng,\n     *                          Ví dụ 2000-1-1\n     *                      ",
                                        "type": "string",
                                        "format": "date",
                                        "example": "2000-01-01"
                                    },
                                    "description": {
                                        "description": "Mô tả chi tiết về bản thân",
                                        "type": "string",
                                        "example": "Mô tả chi tiết về bản thân"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Profile Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Saved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Token is invalid or missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "500": {
                        "description": "Update Profile Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Save failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/profile/upload-avatar": {
            "post": {
                "tags": [
                    "PersonalInfo"
                ],
                "summary": "Upload Avatar",
                "description": "Upload Avatar",
                "operationId": "16164780291b0cc069db4a972d336257",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "properties": {
                                    "avatar": {
                                        "description": "The file to upload",
                                        "type": "string",
                                        "format": "binary"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Update Profile Successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Saved"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated - Token is invalid or missing",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Unauthenticated."
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Update Profile Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "message": {
                                            "type": "string",
                                            "example": "Save failed"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "description": "JWT Bearer token",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}